requirements.txt
Flask
Flask-CORS
python-dotenv
PyJWT
Flask-Bcrypt
uuid
gunicorn
psycopg2-binary